Token security (T13):
- short-lived access token + longer-lived, rotating refresh token (jti + type claims)
- POST /auth/refresh (rotation) and POST /auth/logout (revoke)
- per-user revocation via Hazelcast: tokens-valid-after watermark + active-refresh
  store, enforced by a custom OAuth2TokenValidator; type claim blocks refresh-as-access
- kept off the hot path so access-token validation stays stateless
Docker & CI (T10):
- multi-stage Dockerfile (slim, non-root, multi-arch JRE)
- docker-compose running the API against PostgreSQL
- GitHub Actions: mvn verify (Testcontainers IT runs on CI Docker) + image build
Docs (T11):
- fill the README design section: component + sequence diagrams, key-generation
  trade-offs, concurrency/resilience, token model, portability and scalability
- refresh the intro and technology list

+++ src/main/java/com/patken/api/url_shortener/repository/UrlShortenerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UrlShortenerRepository extends JpaRepository<UrlEntity, Integer> {
+public interface UrlShortenerRepository extends JpaRepository<UrlEntity, Long> {
 
     Optional<UrlEntity> findUrlEntityByOriginalUrl(String originalUrl);
